update zn_screen_layout_add() to change screen position
